% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plato}
\alias{plato}
\title{Results of touch type experiment with two subject using
the PlatoWork tDCS headset for possible improvement of
learning speed.}
\format{
A data frame with 225 rows and 6 variables:
\describe{
\item{date}{time of test}
\item{subject}{name of person doing the typing test}
\item{wpm}{words per minute in typing test}
\item{error}{error rate in typing test}
\item{stimulus}{type of stimulus}
\item{session}{session id, unique for each subject}
...
}
}
\usage{
plato
}
\description{
Results of touch type experiment with two subject using
the PlatoWork tDCS headset for possible improvement of
learning speed.
}
\keyword{datasets}
